adjusted jboss subsystem to refactorings in BPM platform

+++ activiti-engine/src/main/java/org/camunda/bpm/container/impl/jmx/deployment/DeployProcessArchiveStep.java
@@ -1,3 +1,15 @@
+/* Licensed under the Apache License, Version 2.0 (the "License");
+ * you may not use this file except in compliance with the License.
+ * You may obtain a copy of the License at
+ * 
+ *      http://www.apache.org/licenses/LICENSE-2.0
+ * 
+ * Unless required by applicable law or agreed to in writing, software
+ * distributed under the License is distributed on an "AS IS" BASIS,
+ * WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+ * See the License for the specific language governing permissions and
+ * limitations under the License.
+ */
 package org.camunda.bpm.container.impl.jmx.deployment;
 
 import java.io.ByteArrayInputStream;
@@ -7,6 +19,8 @@ import java.util.HashMap;
 import java.util.List;
 import java.util.Map;
 import java.util.Map.Entry;
+import java.util.logging.Level;
+import java.util.logging.Logger;
 
 import org.activiti.engine.ActivitiException;
 import org.activiti.engine.ProcessEngine;
@@ -33,9 +47,12 @@ import org.camunda.bpm.engine.application.ProcessApplicationRegistration;
  */
 public class DeployProcessArchiveStep extends MBeanDeploymentOperationStep {
   
+  private final static Logger LOGGER = Logger.getLogger(DeployProcessArchiveStep.class.getName());      
+  
   protected final ProcessArchiveXml processArchive;
   protected URL metaFileUrl;
   protected Deployment deployment;
+  protected ProcessApplicationRegistration registration;
   
   public DeployProcessArchiveStep(ProcessArchiveXml parsedProcessArchive, URL url) {
     processArchive = parsedProcessArchive;
@@ -83,6 +100,8 @@ public class DeployProcessArchiveStep extends MBeanDeploymentOperationStep {
       deploymentMap.putAll(scanner.findResources(processApplicationClassloader, paResourceRoot, metaFileUrl));
     }
     
+    logDeploymentSummary(deploymentMap);
+    
     // perform process engine deployment
     RepositoryService repositoryService = processEngine.getRepositoryService();
     DeploymentBuilder deploymentBuilder = repositoryService.createDeployment();
@@ -103,8 +122,9 @@ public class DeployProcessArchiveStep extends MBeanDeploymentOperationStep {
     // perform the process engine deployment
     deployment = deploymentBuilder.deploy();
     
-    // register the deployment
-    ProcessApplicationRegistration registration = processEngine.getManagementService().activateDeploymentForApplication(deployment.getId(), processApplication.getReference());
+    // register the deployment 
+    // TODO: think about turning the registration into a seperate step.
+    registration = processEngine.getManagementService().activateDeploymentForApplication(deployment.getId(), processApplication.getReference());
     
     // add attachment
     Map<String, ProcessApplicationRegistration> processArchiveDeploymentMap = operationContext.getAttachment(Attachments.PROCESS_ARCHIVE_DEPLOYMENT_MAP);
@@ -115,17 +135,36 @@ public class DeployProcessArchiveStep extends MBeanDeploymentOperationStep {
     processArchiveDeploymentMap.put(processArchive.getName(), registration);    
   }
 
+  protected void logDeploymentSummary(Map<String, byte[]> deploymentMap) {
+    StringBuilder builder = new StringBuilder();
+    builder.append("Deployment summary for process archive '"+processArchive.getName()+"': \n");
+    builder.append("\n");
+    for (String resourceName : deploymentMap.keySet()) {
+      builder.append("        "+resourceName);
+      builder.append("\n");
+    }
+    builder.append("\n");
+    
+    LOGGER.log(Level.INFO, builder.toString());
+  }
+
   public void cancelOperationStep(MBeanDeploymentOperation operationContext) {   
     
     final MBeanServiceContainer serviceContainer = operationContext.getServiceContainer();    
+
+    // if a registration was performed, remove it.
+    if(registration != null) {
+      registration.unregister();
+    }
     
-    // delete deployment if we were able to create one.
-    if(deployment != null) {      
+    // delete deployment if we were able to create one AND if isDeleteUponUndeploy is set.
+    if(deployment != null && PropertyHelper.getBooleanProperty(processArchive.getProperties(), ProcessArchiveXml.PROP_IS_DELETE_UPON_UNDEPLOY, false)) {      
       ProcessEngine processEngine = getProcessEngine(serviceContainer);      
       if(processEngine != null) {
         processEngine.getRepositoryService().deleteDeployment(deployment.getId(), true);
       }
     }
+    
   }
 
   protected ProcessEngine getProcessEngine(final MBeanServiceContainer serviceContainer) {
